Prototype snapshot testing approach. (#11)
* stop using frames for timing since they vary, use time instead

* r-a renamed things a bit weirdly

* Initial prototype of snapshot testing

* More work on the snapshot testing

* Remove absolute pathing from audio as it stops other tools working

* Update tests/snapshot.rs

Co-authored-by: Davi Silva  <blueridanus@users.noreply.github.com>

* Update snapshots for new test format

* Plot the data!

* Fix the sample range adjustment in the test

* end timestamp should be on the first silence frame, shouldnt include redemption_time

* Add in redemption line

* Add back in passing test

* use default VAD thresholds from reference impl

* forgot the snapshots

* plotting script: PEP 723, shebang, save to png

* eh uv is better

* plot the time of transition events

* adjust plot dpi, padding

* more tests

* oops, fixed sample rate on audios

* Fix remainder based error. (#13)

* Fix remainder based error.

So previously the following lines added in a bit of error carried
forwards:

```
let start_chunk = self.processed_samples / vad_segment_length;

// ...

for i in start_chunk..num_chunks {
    let start_idx = i * vad_segment_length;
```

This is because as we eagerly process the audio that comes in we
process the remainder as well. So the start chunk could actually be
somewhere within a "perfect" chunk boundary. Then with integral division
we lose that fractional part and remainders in the audio are processed
>1 time. In the snapshots you can see this as the silence
current_samples line on the plots was a wavy line. This is because each
process call we're processing a different number of samples because
we're also adding the remainder from the previous samples into our call
as well and this increases/decreases as it wraps around a vad frame.

This doesn't really change the snapshot results too much, the silence
frames are now uniform straightness and some start/end times have
shifted by 20ms it seems fairly evident this has fixed an issue given
the uniform length input produces a uniformly increasing line!

* Remove println!

* Fix issue with < segment length chunk sizes!

We should think about the unprocessed remainder

* Make test failing a bit fancier

* Make the printout a bit more aesthetic

* Add some more spacing

* Add some more notes for contributors

* Put the audio in lfs?

---------

Co-authored-by: Davi Silva <davi@emotech.co>
Co-authored-by: Davi Silva <blueridanus@users.noreply.github.com>