Fix panic in `get_current_speech` (#56)
* Fix panic in `get_current_speech`

If `get_current_speech` duration doesn't map to index it would panic but
there may well be audio in the cached storage. This PR adds a check on
if the duration will map to an index and if that will fail, if it will
fail it uses the cached speech.

* Truncate if padding exceeds buffer

* Try to improve testing over pre-speech padding

* Add a missing sat_sub

* Stop segments from possibly overlapping